refactor(app): extract conversation controller
