+++ WISHLIST.md
@@ -33,8 +33,7 @@ Items are removed when completed.
   - Make log rewrites (after truncate/in-place edit) atomic via temp file + rename — [OPEN]
   - Optionally append a log marker indicating manual history edits — [OPEN]
 - Height/scroll DRYing — [PARTIAL]
-  - Prefer `App::calculate_available_height` everywhere; keep a single helper for “scroll selected index into view” (already added) — [PARTIAL]
-  - Standardize usage in renderer and chat loop — [OPEN]
+  - Continue standardizing on the conversation controller helpers across renderer and chat loop — [OPEN]
 - Stream spawning DRYing — [PARTIAL]
   - Use a config struct (already added) and consider moving the helper out of `chat_loop` for reuse — [PARTIAL]
 - Tests — [OPEN]
